Cleanup westcos driver to remove warnings - by François Leblanc.
git-svn-id: https://www.opensc-project.org/svnp/opensc/trunk@3786 c6295689-39f2-0310-b995-f0e70906c6a9

+++ src/libopensc/cardctl.h
@@ -488,8 +488,8 @@ typedef struct sc_cardctl_asepcos_activate_file {
 
 typedef struct {
 	int key_reference;
-	int key_len; //8, 16 or 24
-	u8	key_value[24];
+	size_t key_len; //8, 16 or 24
+	u8 key_value[24];
 }sc_autkey_t;
 
 typedef struct {
